# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeLens API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Memory {
  id: number;
  title: string;
  content: string;
  category: string;
  /** @nullable */
  emotionalTone?: string | null;
  importance: number;
  tags: string[];
  createdAt: string;
}

export interface MemoryInput {
  title: string;
  content: string;
  category: string;
  emotionalTone?: string;
  importance?: number;
  tags?: string[];
}

export interface Decision {
  id: number;
  title: string;
  description: string;
  reasoning: string;
  /** @nullable */
  outcome?: string | null;
  impact: string;
  createdAt: string;
}

export interface DecisionInput {
  title: string;
  description: string;
  reasoning: string;
  outcome?: string;
  impact: string;
}

export interface TimelineEvent {
  id: number;
  title: string;
  description: string;
  eventType: string;
  date: string;
  milestone: boolean;
  createdAt: string;
}

export interface TimelineEventInput {
  title: string;
  description: string;
  eventType: string;
  date: string;
  milestone?: boolean;
}

export interface Insight {
  id: number;
  title: string;
  content: string;
  insightType: string;
  period: string;
  relatedMemories: number[];
  createdAt: string;
}

export interface InsightInput {
  insightType: string;
  period: string;
}

export interface CallRecord {
  id: number;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  summary?: string | null;
  memoriesExtracted: number;
  decisionsExtracted: number;
  status: string;
  createdAt: string;
}

export interface CallInput {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface ConversationInput {
  title: string;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: Message[];
}

export interface DashboardSummary {
  totalMemories: number;
  totalDecisions: number;
  totalTimelineEvents: number;
  totalInsights: number;
  totalCalls: number;
  growthScore: number;
  recentInsights: Insight[];
  recentMemories: Memory[];
}

export interface ActivityItem {
  id: string;
  type: string;
  title: string;
  description: string;
  createdAt: string;
}

export interface GrowthDataPoint {
  month: string;
  memories: number;
  decisions: number;
  insights: number;
}

export type ListMemoriesParams = {
category?: string;
search?: string;
limit?: number;
};

export type ListDecisionsParams = {
search?: string;
};

export type ListTimelineEventsParams = {
type?: string;
};