# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LifeLens API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: memories
    description: Memory operations
  - name: decisions
    description: Decision operations
  - name: timeline
    description: Timeline event operations
  - name: insights
    description: Insight operations
  - name: calls
    description: Call history operations
  - name: chat
    description: AI chat operations
  - name: dashboard
    description: Dashboard summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /memories:
    get:
      operationId: listMemories
      tags: [memories]
      summary: List all memories
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
    post:
      operationId: createMemory
      tags: [memories]
      summary: Create a new memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "201":
          description: Created memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"

  /memories/{id}:
    get:
      operationId: getMemory
      tags: [memories]
      summary: Get a memory by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "404":
          description: Not found
    delete:
      operationId: deleteMemory
      tags: [memories]
      summary: Delete a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /decisions:
    get:
      operationId: listDecisions
      tags: [decisions]
      summary: List all decisions
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of decisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Decision"
    post:
      operationId: createDecision
      tags: [decisions]
      summary: Create a new decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionInput"
      responses:
        "201":
          description: Created decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"

  /decisions/{id}:
    delete:
      operationId: deleteDecision
      tags: [decisions]
      summary: Delete a decision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /timeline:
    get:
      operationId: listTimelineEvents
      tags: [timeline]
      summary: List all timeline events
      parameters:
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
    post:
      operationId: createTimelineEvent
      tags: [timeline]
      summary: Create a new timeline event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimelineEventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineEvent"

  /timeline/{id}:
    delete:
      operationId: deleteTimelineEvent
      tags: [timeline]
      summary: Delete a timeline event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /insights:
    get:
      operationId: listInsights
      tags: [insights]
      summary: List all insights
      responses:
        "200":
          description: List of insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insight"
    post:
      operationId: generateInsight
      tags: [insights]
      summary: Generate a new AI insight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsightInput"
      responses:
        "201":
          description: Generated insight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insight"

  /calls:
    get:
      operationId: listCalls
      tags: [calls]
      summary: List call history
      responses:
        "200":
          description: List of calls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CallRecord"
    post:
      operationId: createCall
      tags: [calls]
      summary: Create/start a call record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallInput"
      responses:
        "201":
          description: Created call record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallRecord"

  /calls/{id}:
    get:
      operationId: getCall
      tags: [calls]
      summary: Get a call record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Call record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallRecord"
        "404":
          description: Not found

  /chat/conversations:
    get:
      operationId: listConversations
      tags: [chat]
      summary: List chat conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [chat]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /chat/conversations/{id}:
    get:
      operationId: getConversation
      tags: [chat]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationWithMessages"
        "404":
          description: Not found
    delete:
      operationId: deleteConversation
      tags: [chat]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /chat/conversations/{id}/messages:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message and stream AI response (SSE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: SSE stream of AI response
          content:
            text/event-stream: {}

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /dashboard/growth:
    get:
      operationId: getGrowthData
      tags: [dashboard]
      summary: Get growth trend data for charts
      responses:
        "200":
          description: Growth data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GrowthDataPoint"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Memory:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        emotionalTone:
          type: ["string", "null"]
        importance:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - category
        - importance
        - tags
        - createdAt

    MemoryInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        emotionalTone:
          type: string
        importance:
          type: integer
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - content
        - category

    Decision:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        reasoning:
          type: string
        outcome:
          type: ["string", "null"]
        impact:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - reasoning
        - impact
        - createdAt

    DecisionInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        reasoning:
          type: string
        outcome:
          type: string
        impact:
          type: string
      required:
        - title
        - description
        - reasoning
        - impact

    TimelineEvent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        eventType:
          type: string
        date:
          type: string
          format: date-time
        milestone:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - eventType
        - date
        - milestone
        - createdAt

    TimelineEventInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        eventType:
          type: string
        date:
          type: string
          format: date-time
        milestone:
          type: boolean
      required:
        - title
        - description
        - eventType
        - date

    Insight:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        insightType:
          type: string
        period:
          type: string
        relatedMemories:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - insightType
        - period
        - relatedMemories
        - createdAt

    InsightInput:
      type: object
      properties:
        insightType:
          type: string
        period:
          type: string
      required:
        - insightType
        - period

    CallRecord:
      type: object
      properties:
        id:
          type: integer
        duration:
          type: ["integer", "null"]
        transcript:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        memoriesExtracted:
          type: integer
        decisionsExtracted:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - memoriesExtracted
        - decisionsExtracted
        - status
        - createdAt

    CallInput:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Conversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    ConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    MessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    ConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required:
        - id
        - title
        - createdAt
        - messages

    DashboardSummary:
      type: object
      properties:
        totalMemories:
          type: integer
        totalDecisions:
          type: integer
        totalTimelineEvents:
          type: integer
        totalInsights:
          type: integer
        totalCalls:
          type: integer
        growthScore:
          type: integer
        recentInsights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"
        recentMemories:
          type: array
          items:
            $ref: "#/components/schemas/Memory"
      required:
        - totalMemories
        - totalDecisions
        - totalTimelineEvents
        - totalInsights
        - totalCalls
        - growthScore
        - recentInsights
        - recentMemories

    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - title
        - description
        - createdAt

    GrowthDataPoint:
      type: object
      properties:
        month:
          type: string
        memories:
          type: integer
        decisions:
          type: integer
        insights:
          type: integer
      required:
        - month
        - memories
        - decisions
        - insights