# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeLens API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: string;
  type: string;
  title: string;
  description: string;
  createdAt: Date;
}