# lib/api-zod/src/generated/types/callRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeLens API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CallRecord {
  id: number;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  summary?: string | null;
  memoriesExtracted: number;
  decisionsExtracted: number;
  status: string;
  createdAt: Date;
}