# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeLens API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Insight } from './insight';
import type { Memory } from './memory';

export interface DashboardSummary {
  totalMemories: number;
  totalDecisions: number;
  totalTimelineEvents: number;
  totalInsights: number;
  totalCalls: number;
  growthScore: number;
  recentInsights: Insight[];
  recentMemories: Memory[];
}