# lib/api-zod/src/generated/types/growthDataPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifeLens API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GrowthDataPoint {
  month: string;
  memories: number;
  decisions: number;
  insights: number;
}